backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
qrcode==7.4.2
pillow>=10.4.0
bcrypt==4.2.0
pydantic[email]
openai>=1.0.0
